requirements.txt
streamlit
youtube-transcript-api
google-generativeai
langchain
python-dotenv

-e .